requirements.txt
javalang>=0.13.0
tqdm>=4.66.0
pyyaml>=6.0.0
datasets>=2.20.0
transformers>=4.44.0
torch>=2.2.0
accelerate>=0.33.0
evaluate>=0.4.1
sentencepiece>=0.1.99